Issue #116: Develop OS Command Extension
* Split OsCommandHelper into OsCommandHelper that was moved back to the MetricsHub engine since it's used by metricshub-wmi-extension and metricshub-winrm-extension. This also led to the creation of  OsCommandService which is responsible for creating embeddedFiles, and pre-processing the Os Command queries before executing them by the OsCommandRequestExecutor.
* Refactored tests to mock the right classes (OsCommandService or OsCommandHelper).
* Corrected a tiny CLI display bug
