+++ metricshub-engine/src/main/java/org/sentrysoftware/metricshub/engine/strategy/utils/OsCommandResult.java
@@ -2,7 +2,7 @@ package org.sentrysoftware.metricshub.engine.strategy.utils;
 
 /*-
  * ╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲
- * MetricsHub OsCommand Extension
+ * MetricsHub MetricsHub Engine
  * ჻჻჻჻჻჻
  * Copyright 2023 - 2024 Sentry Software
  * ჻჻჻჻჻჻
