+++ metricshub-oscommand-extension/pom.xml
@@ -150,7 +150,7 @@
 										<limit>
 											<counter>COMPLEXITY</counter>
 											<value>COVEREDRATIO</value>
-											<minimum>0.55</minimum>
+											<minimum>0.52</minimum>
 										</limit>
 									</limits>
 								</rule>
